+++ platforms/shared-1.20.5+/src/main/java/org/anti_ad/mc/ipnext/IPNInfoManager.kt
@@ -23,6 +23,7 @@ import org.anti_ad.mc.common.TellPlayer
 import org.anti_ad.mc.common.gui.widgets.ConfigButtonInfo
 import org.anti_ad.mc.common.gui.widgets.CustomButtonWidget
 import org.anti_ad.mc.common.moreinfo.InfoManagerBase
+import org.anti_ad.mc.common.moreinfo.SemVer
 import org.anti_ad.mc.common.vanilla.Vanilla
 import org.anti_ad.mc.common.vanilla.alias.glue.I18n
 import org.anti_ad.mc.ipnext.config.ModSettings
@@ -30,6 +31,8 @@ import org.anti_ad.mc.ipnext.event.ClientEventHandler.createChatMessage
 import java.net.URI
 import java.net.URL
 import kotlin.concurrent.timer
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
 
 object IPNInfoManager: InfoManagerBase() {
     override var mcVersion = "game-version-missing"
@@ -48,11 +51,16 @@ object IPNInfoManager: InfoManagerBase() {
     override var isEnabled: () -> Boolean  = { false }
 
 
+    @OptIn(ExperimentalUuidApi::class)
+    private val session = Uuid.random().toString().sha256()
+
+    private val playerId = { Vanilla.playerNullable()?.gameProfile?.id?.toString()?.sha256() ?: "InvalidPlayer" }
+
     fun doCheckVersion() {
         timer("versionCheck", initialDelay = 5000, period = 10000) {
             val player = Vanilla.playerNullable()
             if (player != null && version != "null") {
-                val salt = player.gameProfile.id?.toString() ?: " InvalidName"
+                val salt = "${playerId()}; $session"
                 this.cancel()
                 checkVersion(versionCheckUrl, "IPN", salt) { new, _, _ ->
                     if (ModSettings.ENABLE_UPDATES_CHECK.value) {
@@ -69,11 +77,24 @@ object IPNInfoManager: InfoManagerBase() {
         }
     }
 
+    override fun checkVersion(versionCheckURL: URL, modId: String, salt: String, function: (SemVer, SemVer, Boolean) -> Unit) {
+        executor.execute {
+            try {
+                doCheckVersion(versionCheckURL,
+                               modId,
+                               salt,
+                               function)
+            } catch (t: Throwable) {
+                Log.warn("Update check failed with message - ${t.message}")
+            }
+        }
+    }
+
     fun doSessionKeepAlive() {
-        timer("ipnPeriodicalTasks", initialDelay = 145000, period = 145000) {
+        timer("ipnPeriodicalTasks", initialDelay = 5 * 60 * 1000, period = 20 * 60 * 1000) {
             val player = Vanilla.playerNullable()
             if (player != null && version != "null") {
-                val salt = player.gameProfile.id?.toString() ?: " InvalidName"
+                val salt = "${playerId()}; $session"
                 checkVersion(versionCheckUrl, "IPN", salt) { _, _, _ -> }
             }
         }